Make show inventory parsing work on a C4500X
In order to make the parser for show inventory work on a Catalyst 4500X,
define a structure within the schema to represent the replaceable SFPs
within the fixed chassis. The component housing the SFPs that might have
a distinct PID for other platforms will simply be called builtin. We
also have to make sure to correctly match the SFPs that are named after
their respective interface.